% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttestRox.R
\name{myttest}
\alias{myttest}
\title{t-test final}
\usage{
myttest(x, y, alpha, paired = FALSE)
}
\arguments{
\item{x}{X Vector (numeric)}

\item{y}{Y vector (numeric)}

\item{alpha}{alpha/error value (numeric)}

\item{paired}{Is sample paired? (default FALSE)}
}
\value{
t-test and input results
}
\description{
t-test final
}
\examples{
\dontrun{myttest(x=rnorm(30,5,2), y=rnorm(30,3,2), alpha=0.05)}
}
