% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRox.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{Plot Rttest}
\usage{
\method{plot}{Rttest}(funObject, i = 5, j = 6, k = 7, l = 8)
}
\arguments{
\item{i}{x vector from object index}

\item{j}{y vector from object index}

\item{k}{TRUE/FALSE independent value from object index}

\item{l}{TRUE/FALSE equal variance value from object index}
}
\value{
dot plot of x and y with test type as the title
}
\description{
Plot Rttest
}
\examples{
\dontrun{plot(funObject)}
}
