% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printRox.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Print method for class Rttest}
\usage{
\method{print}{Rttest}(funObject, ...)
}
\arguments{
\item{funObject}{object made when myttest() is run}

\item{...}{Additional arguments as needed}
}
\value{
a table with the data set, the confidence interval, and hypothesis choice
}
\description{
Print method for class Rttest
}
\examples{
\dontrun{print(funObject)}
}
\author{
Jennifer Vance
}
