% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyOutput.R
\name{shinyOutput}
\alias{shinyOutput}
\title{DynamicShiny}
\usage{
shinyOutput()
}
\value{
Plots CI
}
\description{
Use widgets to plot and run ttest on x and y.
}
\section{shiny}{
 The shiny web server will be used.
}

\examples{
\dontrun{shinyOutput()}
}
