% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRox.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{Plot Rttest}
\usage{
\method{plot}{Rttest}(funObject, i = 5, j = 6, k = 3, l = 7)
}
\arguments{
\item{funObject}{name used when function is run}

\item{i}{x vector from object index}

\item{j}{y vector from object index}

\item{k}{The test type (i.e. paired, t-test, or Welch)}

\item{l}{TRUE/FALSE value for paired}
}
\value{
box plot of x and y with test type as the title
}
\description{
Plot Rttest
}
\examples{
\dontrun{plot(funObject)}
}
