% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyOutput.R
\name{shinyOutput}
\alias{shinyOutput}
\title{Dynamic output using shiny}
\usage{
shinyOutput()
}
\value{
Plots dataframe CI
}
\description{
Dynamic output using shiny
}
\section{shiny}{
 The shiny web server will be used.
}

\examples{
\dontrun{shinyOutput()}
}
